build: Disable git integration, remedy binding
Git integration was causing a failed build with the 5.0.1 ebuild since
the git repository is understandably missing in the tarball. In
addition, on systems where gtk#-3 is also installed the build
experiences binding version mismatches (particularly with gdk-sharp and
glib-sharp).

This patch removes the git execution and sets specific_version for
libraries with judicial use of sed.