dev-dotnet/dotnetcore-sdk-bin: fix 5.0-pre7
dotnetcore-sdk-bin-common was not updated to pre7 version, meaning it
couldn't work. Also arm and arm64 versions were not added to the common
package making it unusable on those archs. If proper support for arm or
arm64 is needed, open an issue or PR with both of these packages,
preferably on all LTS versions.

Signed-off-by: Gergely Nagy <ngg@ngg.hu>